+++ ram-back/src/__tests__/sale.test.ts
@@ -1,7 +1,5 @@
 // (c) Delta Software 2023, rights reserved.
 
-// import request from "supertest";
-// import { app } from "../controller";
 import { createSale } from "../app/sale";
 import { getDataSource } from "../arch/db-client";
 import { createAssuranceType } from "../app/assuranceType";
@@ -28,19 +26,16 @@ describe("sale", () => {
       const { sale, error } = await createSale({
         policyNumber: "123456789",
         assuranceType: {
-          name: "test-assurance-type-1",
-          description: "test-assurance-type-1-description",
           id: "test-at-1",
         },
         user: {
-          email: "test@delta.tec.mx",
-          password: "test-password",
           id: "test-user",
         },
         sellDate: new Date("2021/01/01"),
         amountInCents: "1234567",
         clientName: "Juan Perez",
         evidenceUrl: "https://www.google.com",
+        periodicity: "mensual",
       });
 
       expect(error).toBeUndefined();
