+++ ram-front/src/__test__/saleshistory.test.tsx
@@ -1,6 +1,6 @@
 // (c) Delta Software 2023, rights reserved.
 
-import { render, screen, act, fireEvent } from "@testing-library/react";
+import { act } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
 import SalesHistory from "../pages/SalesHistory";
 import { RecoilRoot } from "recoil";
