+++ ram-front/src/pages/Home.tsx
@@ -51,6 +51,14 @@ export default function Home() {
           </a>
         </li>
         <li>
+          <a
+            href="/sales-history"
+            className="underline text-red-500 hover:text-red-800"
+          >
+            all sales
+          </a>
+        </li>
+        <li>
           <button
             onClick={() => {
               logout();
